consolidate tui constants
