--- src/tui/constants/layout.js
@@ -1,16 +1,0 @@
-export const LAYOUT = {
-  /** Height reserved for modal headers */
-  HEADER_HEIGHT: 4,
-  /** Margin used for borders in modal layouts */
-  BORDER_MARGIN: 2,
-  /** Number of characters used for ellipsis when truncating */
-  ELLIPSIS_LENGTH: 3,
-  /** Maximum width for separator lines */
-  MAX_SEPARATOR_WIDTH: 60,
-  /** Minimum width expected for terminal content */
-  MIN_TERMINAL_WIDTH: 40,
-  /** Fallback terminal height when detection fails */
-  DEFAULT_TERMINAL_HEIGHT: 24,
-  /** Fallback terminal width when detection fails */
-  DEFAULT_TERMINAL_WIDTH: 80,
-};
