+++ src/tui/event-decorators/command.js
@@ -2,7 +2,7 @@ import { formatTimestamp } from '../ui-utils/entry-utils.js';
 import { TextLayout } from '../ui-utils/text-layout.js';
 import { sanitizeText } from '../ui-utils/line-formatter.js';
 
-import { TRACE_ICONS } from '../constants/ui.js';
+import { TRACE_ICONS } from '../constants/index.js';
 
 /** @type {import('./index.js').EventDecorator} */
 export const command = {
