+++ src/tui/event-decorators/unknown.js
@@ -1,7 +1,7 @@
 import { formatTimestamp } from '../ui-utils/entry-utils.js';
 import { TextLayout } from '../ui-utils/text-layout.js';
 
-import { TRACE_ICONS } from '../constants/ui.js';
+import { TRACE_ICONS } from '../constants/index.js';
 
 /** @type {import('./index.js').EventDecorator} */
 export const unknown = {
