+++ src/tui/event-decorators/writeFile.js
@@ -2,7 +2,7 @@ import { formatTimestamp } from '../ui-utils/entry-utils.js';
 import { TextLayout } from '../ui-utils/text-layout.js';
 import { sanitizeText } from '../ui-utils/line-formatter.js';
 
-import { TRACE_ICONS, TRACE_COLORS } from '../constants/ui.js';
+import { TRACE_ICONS, TRACE_COLORS } from '../constants/index.js';
 
 /** @type {import('./index.js').EventDecorator} */
 export const writeFile = {
