+++ src/tui/views/log-viewer/LogViewer.js
@@ -8,7 +8,7 @@ import { AppContainer } from '../AppContainer.js';
 import { LineRenderer } from './LineRenderer.js';
 import { useStdoutDimensions } from '../../hooks/useStdoutDimensions.js';
 import { useSelection } from '../../hooks/useSelection.js';
-import { SHORTCUTS, buildFooter } from '../../constants/shortcuts.js';
+import { SHORTCUTS, buildFooter } from '../../constants/index.js';
 import { isEnterKey } from '../../ui-utils/input-utils.js';
 
 const Entry = ({ item, selected }) =>
