+++ src/tui/views/trace-details/TraceDetailsView.js
@@ -4,7 +4,7 @@ import { LineRenderer } from '../log-viewer/LineRenderer.js';
 import { AppContainer } from '../AppContainer.js';
 import { useStdoutDimensions } from '../../hooks/useStdoutDimensions.js';
 import { buildEntryLines } from '../../ui-utils/entry-utils.js';
-import { SHORTCUTS, buildFooter } from '../../constants/shortcuts.js';
+import { SHORTCUTS, buildFooter } from '../../constants/index.js';
 import { isExitKey } from '../../ui-utils/input-utils.js';
 
 export const TraceDetailsView = ({
