+++ src/tui/views/trace-types-filter/TraceTypesFilterView.js
@@ -4,7 +4,7 @@ import { MultiSelect } from '@inkjs/ui';
 import { useStdoutDimensions } from '../../hooks/useStdoutDimensions.js';
 import { AppContainer } from '../AppContainer.js';
 import { DEFAULT_FILTERS } from '../../ui-utils/entry-utils.js';
-import { SHORTCUTS, buildFooter } from '../../constants/shortcuts.js';
+import { SHORTCUTS, buildFooter } from '../../constants/index.js';
 import { isBackKey } from '../../ui-utils/input-utils.js';
 
 /**
